requirements.txt
streamlit
openai-whisper
torch==2.0.0
